[Fleet] Filter hosted policies from menu (#97380)
## Summary

 * 4C & 7C from https://github.com/elastic/kibana/issues/91906 about hiding hosted agent policy from select menus
 * Also changed reassign agent to new policy UI from flyout to modal as requested
<details><summary>Screenshots</summary>

<h3>Current</h3>
<img width="1090" alt="Screen Shot 2021-04-16 at 1 37 02 PM" src="https://user-images.githubusercontent.com/57655/115075124-9d9ea600-9ec8-11eb-8e88-d8f14f62613f.png">


<h3>PR</h3>
<img width="1092" alt="Screen Shot 2021-04-16 at 1 43 51 PM" src="https://user-images.githubusercontent.com/57655/115075127-9d9ea600-9ec8-11eb-8c47-c3b1fbeb96eb.png">

<h3>current</h3>
<img width="1112" alt="Screen Shot 2021-04-16 at 1 44 41 PM" src="https://user-images.githubusercontent.com/57655/115075129-9e373c80-9ec8-11eb-9f93-cac511c9d53d.png">

<h3>PR</h3>
<img width="1106" alt="Screen Shot 2021-04-16 at 1 50 28 PM" src="https://user-images.githubusercontent.com/57655/115075130-9e373c80-9ec8-11eb-93fb-0adfa5ae81ba.png">


<h3>current</h3>
<img width="1106" alt="Screen Shot 2021-04-16 at 1 35 48 PM" src="https://user-images.githubusercontent.com/57655/115075121-9d9ea600-9ec8-11eb-8605-89336371938b.png">

<h3>PR</h3>
<img width="1106" alt="Screen Shot 2021-04-16 at 2 35 15 PM" src="https://user-images.githubusercontent.com/57655/115075133-9ecfd300-9ec8-11eb-9458-0b7f9cf76c10.png">

<img width="2136" alt="Screen Shot 2021-04-16 at 3 24 16 PM" src="https://user-images.githubusercontent.com/57655/115075141-a0010000-9ec8-11eb-896c-efc4be4be5ac.png">

</details>

### Checklist
- [x] Any text added follows [EUI's writing guidelines](https://elastic.github.io/eui/#/guidelines/writing), uses sentence case text and includes [i18n support](https://github.com/elastic/kibana/blob/master/packages/kbn-i18n/README.md)
