+++ x-pack/plugins/fleet/public/applications/fleet/components/package_icon.tsx
@@ -14,7 +14,18 @@ import { usePackageIconType } from '../hooks';
 
 export const PackageIcon: React.FunctionComponent<
   UsePackageIconType & Omit<EuiIconProps, 'type'>
-> = ({ packageName, version, icons, tryApi, ...euiIconProps }) => {
+> = ({ size = 's', packageName, version, icons, tryApi, ...euiIconProps }) => {
   const iconType = usePackageIconType({ packageName, version, icons, tryApi });
-  return <EuiIcon size="s" type={iconType} {...euiIconProps} />;
+  return (
+    <EuiIcon
+      // when a custom SVG is used the logo is rendered with <img class="euiIcon euiIcon--small">
+      // this collides with some EuiText (+img) CSS from the EuiIcon component
+      // which  makes the button large, wide, and poorly layed out
+      // override those styles until the bug is fixed or we find a better approach
+      style={{ margin: 'unset', width: 'unset' }}
+      size={size}
+      type={iconType}
+      {...euiIconProps}
+    />
+  );
 };
