+++ x-pack/plugins/fleet/public/applications/fleet/sections/agents/components/index.tsx
@@ -6,7 +6,7 @@
  */
 
 export * from './loading';
-export * from './agent_reassign_policy_flyout';
+export * from './agent_reassign_policy_modal';
 export * from './agent_enrollment_flyout';
 export * from './agent_health';
 export * from './agent_unenroll_modal';
